+++ extensions/void/src/ai/vectorStore.ts
@@ -2,16 +2,19 @@ import { Client as OpenSearchClient } from "@opensearch-project/opensearch";
 import { Document } from "@langchain/core/documents";
 import { OpenSearchVectorStore } from "@langchain/community/vectorstores/opensearch";
 import { EmbeddingsInterface } from "@langchain/core/embeddings";
+import { ApiConfig, getApiConfig, VectorStore } from "../config";
 
 const INDEX_NAME = "void";
 
-// TODO other vector store clients
-const getVectorStoreClient = async (embeddingApi: EmbeddingsInterface) => {
-	const client = new OpenSearchClient({
-		nodes: ["http://127.0.0.1:9200"],
-	});
-
-	return client;
+const getVectorStoreClient = async (apiConfig: ApiConfig) => {
+	switch (apiConfig.vectorStore) {
+		case VectorStore.OPENSEARCH:
+			return new OpenSearchClient({
+				nodes: [apiConfig.openSearch.endpoint],
+			});
+		default:
+			return null;
+	}
 };
 
 const uploadDocuments = async (
@@ -66,13 +69,18 @@ const getStoredMtime = async (
 };
 
 export default async (embeddingApi: any) => {
-	const client = await getVectorStoreClient(embeddingApi);
+	const apiConfig = getApiConfig();
+	const client = await getVectorStoreClient(apiConfig);
 
-	return {
-		uploadDocuments: (documents: Document[]) =>
-			uploadDocuments(client, embeddingApi, documents),
-		deleteDocuments: (path: string) => deleteDocuments(client, path),
-		getStoredMtime: (path: string) =>
-			getStoredMtime(client, embeddingApi, path),
-	};
+	if (client) {
+		return {
+			uploadDocuments: (documents: Document[]) =>
+				uploadDocuments(client, embeddingApi, documents),
+			deleteDocuments: (path: string) => deleteDocuments(client, path),
+			getStoredMtime: (path: string) =>
+				getStoredMtime(client, embeddingApi, path),
+		};
+	} else {
+		return null;
+	}
 };
