+++ extensions/void/src/shared_types.ts
@@ -1,6 +1,6 @@
 
 import * as vscode from 'vscode';
-import { ApiConfig } from './common/sendLLMMessage';
+import { ApiConfig } from './config';
 
 // a selection is a frozen snapshot
 type Selection = { selectionStr: string, selectionRange: vscode.Range, filePath: vscode.Uri }
